requirements.txt
anyascii==0.3.2
anyio==4.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.6
django-filter==24.2
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
editables==0.5
et-xmlfile==1.1.0
filelock==3.13.4
filetype==1.2.0
h11==0.14.0
hatch==1.9.4
hatchling==1.21.1
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
keyring==25.1.0
l18n==2021.3
laces==0.1.1
lxml==5.1.0
markdown-it-py==3.0.0
marshmallow==3.21.2
mdurl==0.1.2
more-itertools==10.2.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pillow_heif==0.16.0
platformdirs==4.2.0
pluggy==1.4.0
ptyprocess==0.7.0
pygltflib==1.16.2
Pygments==2.17.2
python-pptx==0.6.23
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.32.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
tomli_w==1.0.0
tomlkit==0.12.4
trove-classifiers==2024.4.10
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.2
virtualenv==20.25.1
wagtail==6.1.1
Willow==1.8.0
wrapt==1.16.0
XlsxWriter==3.1.9
zstandard==0.22.0
